feat(repo-explorer): added pass-through links for entries
