+++ src/features/repo-explorer/components/entries-view/index.scss
@@ -53,8 +53,9 @@
             padding: 4px;
         }
 
-        span {
+        .origin-link {
             margin-left: 12px;
+            color: var(--clr-text);
         }
     }
 }
