+++ src/features/repo-explorer/components/entries-view/index.tsx
@@ -3,6 +3,9 @@ import dayjs from "dayjs";
 import { List } from "antd";
 import cn from "classnames";
 import { Link } from "react-router-dom";
+import { RepoIdentity } from "../../../../models";
+import { GITHUB_DOMAIN } from "../../../../shared/get-env";
+import { useBranch } from "../../hooks";
 import SkeletonArea from "../skeleton-area";
 
 // FIXME: import as ReactComponent
@@ -12,10 +15,12 @@ import logo from "./placeholder.png";
 import "./index.scss";
 
 type GitFile = { type: string; name: string };
+type GitFileViewProps = GitFile & { repo: RepoIdentity; branch: string };
 type Props = {
     loading?: boolean;
     files: Array<GitFile>;
     className?: string;
+    repo: RepoIdentity;
     lastCommit?: {
         message: string;
         login?: string;
@@ -25,7 +30,8 @@ type Props = {
     };
 };
 
-const EntriesView = ({ loading, files, lastCommit, className }: Props) => {
+const EntriesView = ({ loading, files, lastCommit, className, repo }: Props) => {
+    const { branch } = useBranch(repo);
     return (
         <div className={cn("repo-git-view", className)}>
             {loading && <SkeletonArea />}
@@ -33,21 +39,27 @@ const EntriesView = ({ loading, files, lastCommit, className }: Props) => {
                 <List
                     header={lastCommit && <LastCommitHeader lastCommit={lastCommit} />}
                     dataSource={files}
-                    renderItem={GitFileView}
+                    renderItem={(item) => <GitFileView {...item} repo={repo} branch={branch} />}
                 />
             )}
         </div>
     );
 };
 
-const GitFileView = ({ name, type }: GitFile) => (
-    <List.Item className="repo-git-view__item">
-        <div>
-            <img alt="type" src={type === "tree" ? FolderIcon : FileIcon} />
-            <span>{name}</span>
-        </div>
-    </List.Item>
-);
+const GitFileView = ({ name, type, repo, branch }: GitFileViewProps) => {
+    const baseUrl = `${GITHUB_DOMAIN}${repo.owner}/${repo.name}`;
+    const link = `${baseUrl}/${type}/${branch ? branch + "/" : ""}${name}`;
+    return (
+        <List.Item className="repo-git-view__item">
+            <div>
+                <img alt="type" src={type === "tree" ? FolderIcon : FileIcon} />
+                <a href={link} target="_blank" rel="noopener noreferrer" className="origin-link">
+                    {name}
+                </a>
+            </div>
+        </List.Item>
+    );
+};
 
 const LastCommitHeader = ({ lastCommit }: { lastCommit: Props["lastCommit"] }) => (
     <div className="repo-git-view__last-commit">
